make qr scanning work
Fixed microphone permissions
Added default page if no permissions
