+++ navigation/Navigation.tsx
@@ -6,7 +6,7 @@ import { FontAwesome } from "@expo/vector-icons";
 import { View } from "react-native";
 import Home from "../screens/Home";
 import Events from "../screens/Events";
-import Camera from "../screens/Camera";
+import CameraScanner from "../screens/CameraScanner";
 import Shop from "../screens/Shop";
 import Profile from "../screens/Profile";
 
@@ -89,7 +89,7 @@ const AppNavigator: React.FC = () => {
       />
       <Tab.Screen
         name="Camera"
-        component={Camera}
+        component={CameraScanner}
         options={{ tabBarLabel: () => null }}
       />
       <Tab.Screen
